fix docker setup
The docker setup on Fedora is getting very messy. Docker is packaged via
the `moby-engine` package, but Docker currently doesn't supports cgroups
v2. So we have to manually switch pack to cgroups v1 by setting a grub
option (see https://blog.christophersmart.com/2019/12/15/enabling-docker-in-fedora-31-by-reverting-to-cgroups-v1/ ).

Then, in order to use docker swarm, we should use docker. Fedora tends
to discourage the use of docker to instead use podman (and
podman-compose), but podman only supports the k8s orchestrator.

Finally, fedora comes with SELinux enabled by default but this is a PITA
for docker usage on a dev machine (aka workstation).
Enabling SELinux rules for bind mounts and entrypoints usages for each
containers of each services on eack kernel version is simply not doable
for a such fast moving environment.
So, for simplicity, we can consider to disable SELinux on this workstation,
untill a better solution is found.