stick to mixxx 2.2.4
As the 2.3.0 update has a weird behavior when pressing the CUE button
while using DVS.

When playing w/ DVS and pressing the cue button to go back to the cue
point, mixxx continues the playback for some ms after going back to the
cuepoint, which moves the cursor ahead, and also continues the playback
for a small amount of time.

As this wierd behavior wasn't here on v2.2.4, we stick to this stable
version.