install ardour via flatpak
As it is able to get sound from external soundcard.
I wasn't able to get (or send) sound anymore to the external soundcard
when using Ardour from the Fedora 34 package manager (dnf). Maybe there
is some Pipewire mess as this version has introduced it.
Using the containerized app w/ flatpak solves this issue at least, but
it could be better to investigate it a bit more in order to use standard
host installation (would be nice to be able to use sound effects installed
by the dnf package manager).