create a unique role to manage flatpak apps
As more and more apps are installed simply with flatpak, group them in a
single role with array vars to specify which app to install or remove,
instead of having one role per app with a flag var to tell whether to
install it or not.

Apps are installed via flatpak instead of dnf as it provides better
(more stable) runtimes, and also more up to date versions.